Move ConfigMap to v1 API out of extensions
