+++ kubernetes-mock/src/main/java/io/fabric8/kubernetes/client/mock/ExtensionsAPIGroupMockClient.java
@@ -15,8 +15,6 @@
  */
 package io.fabric8.kubernetes.client.mock;
 
-import io.fabric8.kubernetes.api.model.extensions.ConfigMap;
-import io.fabric8.kubernetes.api.model.extensions.ConfigMapList;
 import io.fabric8.kubernetes.api.model.extensions.DaemonSet;
 import io.fabric8.kubernetes.api.model.extensions.DaemonSetList;
 import io.fabric8.kubernetes.api.model.extensions.Deployment;
@@ -34,7 +32,6 @@ import io.fabric8.kubernetes.client.mock.impl.MockDeployment;
 import io.fabric8.kubernetes.client.mock.impl.MockIngress;
 import io.fabric8.kubernetes.client.mock.impl.MockJobs;
 import io.fabric8.kubernetes.client.mock.impl.MockThirdPartyResource;
-import io.fabric8.kubernetes.client.mock.impl.donable.MockDoneableConfigMap;
 import io.fabric8.kubernetes.client.mock.impl.donable.MockDoneableDaemonSet;
 import io.fabric8.kubernetes.client.mock.impl.donable.MockDoneableDeployment;
 import io.fabric8.kubernetes.client.mock.impl.donable.MockDoneableIngress;
@@ -53,7 +50,6 @@ public class ExtensionsAPIGroupMockClient implements Replayable<ExtensionsAPIGro
     private MockDeployment deployments = new MockDeployment();
     private MockIngress ingresses = new MockIngress();
     private MockDaemonSet daemonSets = new MockDaemonSet();
-    private MockConfigMap configMaps = new MockConfigMap();
     private MockThirdPartyResource thirdPartyResources = new MockThirdPartyResource();
 
     public ExtensionsAPIGroupMockClient() {
@@ -62,7 +58,6 @@ public class ExtensionsAPIGroupMockClient implements Replayable<ExtensionsAPIGro
         expect(client.ingress()).andReturn(ingresses.getDelegate()).anyTimes();
         expect(client.ingresses()).andReturn(ingresses.getDelegate()).anyTimes();
         expect(client.daemonSets()).andReturn(daemonSets.getDelegate()).anyTimes();
-        expect(client.configMaps()).andReturn(configMaps.getDelegate()).anyTimes();
         expect(client.thirdPartyResources()).andReturn(thirdPartyResources.getDelegate()).anyTimes();
         client.close();
         EasyMock.expectLastCall().anyTimes();
@@ -81,7 +76,6 @@ public class ExtensionsAPIGroupMockClient implements Replayable<ExtensionsAPIGro
         deployments.verify();
         ingresses.verify();
         daemonSets.verify();
-        configMaps.verify();
         thirdPartyResources.verify();
         EasyMock.verify(client);
     }
@@ -105,10 +99,6 @@ public class ExtensionsAPIGroupMockClient implements Replayable<ExtensionsAPIGro
 
     public MockOperation<DaemonSet, DaemonSetList, MockDoneableDaemonSet, MockResource<DaemonSet, MockDoneableDaemonSet, Boolean>> daemonSets() {
         return daemonSets;
-    }
-
-    public MockOperation<ConfigMap, ConfigMapList, MockDoneableConfigMap, MockResource<ConfigMap, MockDoneableConfigMap, Boolean>> configMaps() {
-        return configMaps;
     }
 
     public MockOperation<ThirdPartyResource, ThirdPartyResourceList, MockDoneableThirdPartyResource, MockResource<ThirdPartyResource, MockDoneableThirdPartyResource, Boolean>> thirdPartyResources() {
